Enhancement - Modified countdown timer functionality
What
=================
Implemented a way for the countdown deadline to be set in the html to
allow for multiple countdown timers to exist on the website.
Modified the countdown on the hypethon page to count down to the time at
which the event starts, rather than when the BETR sales happen.