Enhancement - Popup system remoddled
What
=================
Remoddled the popup system to be less obstructive when it overlayed the
content of the website.
Added ability to record when user has explicitly closed a specific
popup, disallowing that popup from showing again.

These changes mean that announcement popups now need to be included
after the `<body>` tag and before the content, which is now wrapped in
`<main>` in order for the announcement popup to change the position of
the content of the site.