Bugfix - homepage responsiveness + php code not parsing
What
=================
Homepage responsiveness broke when the long string for the token address
started pushing the right border off the view edge. This is now being
ellipsed accordingly.
Php code in the parameterised variable of the string for the currency
"alt" tags were not parsing due to being surrounded by a "ucwords()"
function.